Add user integration in all APIs

+++ pkg/middlewares/auth.go
@@ -99,7 +99,7 @@ func AttachUser(next echo.HandlerFunc) echo.HandlerFunc {
 			return c.JSON(http.StatusUnauthorized, "Missing email claim")
 		}
 
-		c.Set("user_id", claims["id"].(float64))
+		c.Set("user_id", uint(claims["id"].(float64)))
 
 		return next(c)
 	}
